IMPORTANT: Changed Collisions
Now, only \\ player.objects.ship // and \\ player.objects.shield; if shield.solid // actively check for collisions in their // .update \\ methods.  *All* objects are set for *passive collision* detection with the \\ collider //.  They no longer have \\ .onCollide // methods, and they call the other objects \\ .onCollide // method.  When objects collide with the shield, if they have \\ .typeN == "bullet" // then they will trigger a collision response in the shield.  When any shape collides with the ship, it's \\ .onCollide // method will not be called and it's \\ .hp // will be 0.   If you want collisions to be checked on \\ computer.objects // you can do that, but you still need to have a \\ .onCollide // method, even if it doesn't do anything, ie. \\ .onCollide = function(self, obj) end //.